pyproject.toml
[project]
name = "integration-synergy-tracker"
version = "0.1.0"
description = "Integration Synergy Strategy Tracker"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "jinja2>=3.1",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "python-dateutil>=2.9",
  "pandas>=2.2",
  "openpyxl>=3.1",
  "weasyprint>=62.3",
  "typer>=0.12",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "black>=24.8",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
